+++ ga_view/src/w_coordsys.hpp
@@ -16,12 +16,10 @@ class w_Coordsys : public QGraphicsView {
     Q_OBJECT
 
   public:
-
     w_Coordsys(Coordsys* cs, std::vector<Coordsys_model*>& vm, QGraphicsScene* scene,
                QWidget* parent = nullptr);
 
   protected:
-
     void resizeEvent(QResizeEvent* event);
     void paintEvent(QPaintEvent* event);
     void drawBackground(QPainter* qp, const QRectF& rect);
@@ -45,11 +43,10 @@ class w_Coordsys : public QGraphicsView {
     void viewResized();
     void mouseMoved(bool hot, mouse_pos_t mouse_pos);
     void modeChanged(pz_action action, pz_mode mode);
-    void undoChanged(int undo_steps);
+    void undoChanged(size_t undo_steps);
     void scalingChanged(axis_scal xscal, axis_scal yscal);
 
   private:
-
     Coordsys* cs;          // coordinate system
     QGraphicsScene* scene; // graphics scene
