+++ ga_view/src/w_statusbar.hpp
@@ -28,10 +28,10 @@ class w_Statusbar : public QWidget {
     // mouse within (min...max) (true) or not (false) cs area, current position is
     // x, y
     void on_mouseMoved(bool hot, mouse_pos_t mouse_pos);
-    void on_currentModelChanged(int currentModelIndex);
-    void on_maximumModelChanged(int maximumModelIndex);
+    void on_currentModelChanged(size_t currentModelIndex);
+    void on_maximumModelChanged(size_t maximumModelIndex);
     void on_modeChanged(pz_action action, pz_mode mode);
-    void on_undoChanged(int undo_steps);
+    void on_undoChanged(size_t undo_steps);
     void on_labelChanged(std::string label);
     void on_scalingChanged(axis_scal xscal, axis_scal yscal);
 
@@ -48,8 +48,8 @@ class w_Statusbar : public QWidget {
     double m_x, m_y; // mouse position in cs
 
 
-    int m_currentModel{0}; // model index (default: show first step)
-    int m_maximumModel{0}; // maximum index of models
+    size_t m_currentModel{0}; // model index (default: show first step)
+    size_t m_maximumModel{0}; // maximum index of models
 
     // model label
     std::string m_label{};
@@ -62,7 +62,7 @@ class w_Statusbar : public QWidget {
     pz_mode m_mode{pz_mode::x_and_y};
 
     // number of undo steps available
-    int m_undo_steps{0};
+    size_t m_undo_steps{0};
 
     // axis scaling
     axis_scal m_xscaling{axis_scal::linear};
